[Title] - Added syntax highlighting Logic, Replaced innerHTML's with innerText [Descripttion] - this commit has two main changes 1. added logic for syntax highlighting inside   `src/MusicBlock_v4_Support/runSyntaxHighlighter.js`. 2. replaced all innerHTML's with innerText.
1. Workflow of syntax highlighting.
  - User types a text, handleInputChange inside Tasks folder handles the
   input and replaces the DOM text. Then it emits an event on input box
   named "InputTriggered".
  - Action for "InputTriggered" is added inside
   `/MusicBlock_v4_Support/addingInputEventListener`. inside this file
   we call `runSyntaxHighlighter` from
   `/MusicBlock_v4_Support/runSyntaxHighlighter`, which performs
   syntax highlighting on text and replaces the DOM.
I choose this workflow to keep support for MusicBlock_v4 language more
like a plugin and keep code Editor logic independent of the language.

2. when a user types `<` this symbol innerHTML interprets any text after
that as a HTML element which causes the text to disappear. I should have
used innerText from the beggining but it kind of slipped my mind that
`<` will be interpreted in the wrong way. so I replace all the innerHTML
will innerText and added tags into DOM using appendChild.